# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyBusiness 2.0 API",
        "description": "MyBusiness API documentation",
        "contact": {
            "email": "sgg.2020@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/access_period/index": {
            "get": {
                "tags": [
                    "Access period"
                ],
                "summary": "Get all access periods",
                "operationId": "indexAccessPeriod",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/access_period/{id}": {
            "get": {
                "tags": [
                    "Access period"
                ],
                "summary": "Get access period by ID",
                "operationId": "showAccessPeriod",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "String ID",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "D9B7E73F-9E5B-4884-AF72-1FE5934A6E3A"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/access_period/add": {
            "post": {
                "tags": [
                    "Access period"
                ],
                "summary": "Create access period",
                "operationId": "f7f6a08aaa9e037df7ccfa5f244aac91",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "term_id": {
                                                "type": "uuid"
                                            },
                                            "admin_beginning_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "admin_ending_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "admin_cancel_section_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "depart_beginning_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "depart_ending_date": {
                                                "type": "string",
                                                "format": "date"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "term_id": "e8356132-0e8c-4448-b547-6254094ddd3b",
                                    "admin_beginning_date": "2010/02/02",
                                    "admin_ending_date": "2010/02/02",
                                    "admin_cancel_section_date": "2010/02/02",
                                    "depart_beginning_date": "2010/02/02",
                                    "depart_ending_date": "2010/02/02"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/access_period/update/{id}": {
            "put": {
                "tags": [
                    "Access period"
                ],
                "summary": "Update access period",
                "operationId": "b34aea3dfca56a2530989a254b0e60fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Uuid",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "uuid",
                            "default": "D9B7E73F-9E5B-4884-AF72-1FE5934A6E3A"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "term_id": {
                                                "type": "uuid"
                                            },
                                            "admin_beginning_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "admin_ending_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "admin_cancel_section_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "depart_beginning_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "depart_ending_date": {
                                                "type": "string",
                                                "format": "date"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "term_id": "e8356132-0e8c-4448-b547-6254094ddd3b",
                                    "admin_beginning_date": "2010/02/02",
                                    "admin_ending_date": "2010/02/02",
                                    "admin_cancel_section_date": "2010/02/02",
                                    "depart_beginning_date": "2010/02/02",
                                    "depart_ending_date": "2010/02/02"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/access_period/delete/{id}": {
            "delete": {
                "tags": [
                    "Access period"
                ],
                "summary": "Delete access period by id",
                "operationId": "deleteAccessPeriod",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Uuid",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "0C6A36BA-10E4-438F-BA86-0D5B68A2BB15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}